move dice class API to correct file
